Powershell package Install requires sudo

+++ install.sh
@@ -69,14 +69,18 @@ function create_directories() {
 }
 
 function install_powershell() {
+  # exit if already installed
+  if dpkg -s powershell-lts &>/dev/null; then echo -e " $GREEN - powershell is installed $CLEAR"; return; fi
+
+  echo -e "$GREEN - Installing powershell $CLEAR"
   tmpDir=$(mktemp -d)
   curl -sSL 'https://launchpad.net/ubuntu/+archive/primary/+files/libicu72_72.1-3ubuntu3_amd64.deb' -o "$tmpDir/libicu72_72.1-3ubuntu3_amd64.deb"
-  dpkg -i "$tmpDir"/libicu72_72.1-3ubuntu3_amd64.deb
+  sudo dpkg -i "$tmpDir"/libicu72_72.1-3ubuntu3_amd64.deb
 
   downloadUrl=$(curl -sSL "https://api.github.com/repos/PowerShell/PowerShell/releases/latest" |
     jq -r '[.assets[] | select(.name | endswith("_amd64.deb")) | .browser_download_url][0]')
   curl -sSL "$downloadUrl" -o "$tmpDir/powershell.deb"
-  dpkg -i "$tmpDir"/powershell.deb
+  sudo dpkg -i "$tmpDir"/powershell.deb
 }
 
 function stow_dotfiles() {
@@ -161,8 +165,8 @@ function main_ubuntu() {
   clone_dotfiles
   create_directories
   install_apt_packages
-  install_oh_my_posh
   install_powershell
+  install_oh_my_posh
   install_nanorc_highlighting
   stow_dotfiles
 }
